vm fixes
Fixes for vm:

Correctly setup omero to autostart after postgres, this fixes a problem
where postgres and omero would initialize at the same time and omero
would time out trying to talk to postgres and mark it as failed.

Init script and support files for setting up OMERO.web.  Install and
setup an initscript for OMERO.web.  Additionally a setup script and
control script for nginx (running as the omero user).

Changes to omerovm.sh to support extra files and changes.