Remove cluster intialization step
As pointed out by Petr, the cluster intialization is now done as a post-install
step during the PSQL installation via Homebrew. Since the default cluster
location is cleaned up by the cleanup_homebrew.sh script, this step is now
redundant and should be ommitted from the script (and our documentation).